init/requirements.txt
requests==2.26.0
psycopg2-binary==2.9.3